+++ src/components/screen/util/campaign-icon.jsx
@@ -1,30 +1,38 @@
 import { React, useEffect, useState } from "react";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faExclamationCircle } from "@fortawesome/free-solid-svg-icons";
 import { useDispatch } from "react-redux";
 import PropTypes from "prop-types";
-import idFromUrl from "../util/helpers/id-from-url";
-import calculateIsPublished from "../util/helpers/calculate-is-published";
+import { useTranslation } from "react-i18next";
+import Spinner from "react-bootstrap/Spinner";
+import idFromUrl from "../../util/helpers/id-from-url";
+import calculateIsPublished from "../../util/helpers/calculate-is-published";
 import {
   api,
   useGetV2ScreensByIdCampaignsQuery,
   useGetV2ScreensByIdScreenGroupsQuery,
-} from "../../redux/api/api.generated.ts";
+} from "../../../redux/api/api.generated.ts";
 
 /**
  * An icon to show if the screen has an active campaign.
  *
  * @param {object} props - The props.
  * @param {string} props.id The id of the screen.
+ * @param {number} props.delay Delay the fetch.
  * @returns {object} The campaign icon.
  */
-function CampaignIcon({ id }) {
+function CampaignIcon({ id, delay }) {
+  const { t } = useTranslation("common", { keyPrefix: "campaign-icon" });
   const dispatch = useDispatch();
-  const [isOverriddenByCampaign, setIsOverriddenByCampaign] = useState(false);
+  const [isOverriddenByCampaign, setIsOverriddenByCampaign] = useState(null);
   const [screenCampaignsChecked, setScreenCampaignsChecked] = useState(false);
   const [allCampaigns, setAllCampaigns] = useState([]);
-  const { data: campaigns } = useGetV2ScreensByIdCampaignsQuery({ id });
-  const { data: groups } = useGetV2ScreensByIdScreenGroupsQuery({ id });
+  const [getData, setGetData] = useState(false);
+
+  const { data: campaigns, isLoading } = useGetV2ScreensByIdCampaignsQuery(
+    { id },
+    { skip: !getData }
+  );
+  const { data: groups, isLoading: isLoadingScreenGroups } =
+    useGetV2ScreensByIdScreenGroupsQuery({ id }, { skip: !getData });
 
   useEffect(() => {
     if (campaigns) {
@@ -65,16 +73,43 @@ function CampaignIcon({ id }) {
     }
   }, [allCampaigns]);
 
-  return (
-    <FontAwesomeIcon
-      icon={faExclamationCircle}
-      style={isOverriddenByCampaign ? { color: "red" } : { color: "grey" }}
-    />
-  );
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setGetData(true);
+    }, delay);
+
+    return () => {
+      clearTimeout(timeout);
+    };
+  }, []);
+
+  if (!getData || isLoading || isLoadingScreenGroups) {
+    return (
+      <div style={{ height: "38px" }}>
+        <Spinner
+          as="span"
+          animation="border"
+          size="sm"
+          role="status"
+          aria-hidden="true"
+          className="m-1"
+        />
+      </div>
+    );
+  }
+
+  return isOverriddenByCampaign
+    ? t("overridden-by-campaign")
+    : t("not-overridden-by-campaign");
 }
+
+CampaignIcon.defaultProps = {
+  delay: 1000,
+};
 
 CampaignIcon.propTypes = {
   id: PropTypes.string.isRequired,
+  delay: PropTypes.number,
 };
 
 export default CampaignIcon;