+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#274](https://github.com/os2display/display-admin-client/pull/274)
+  - Added screen status to screen list.
+  - Refactored screen status on screen edit.
+  - Change campaign icon in screen list to boolean text.
+
 ## [2.3.0] - 2025-03-24
 
 - [#279](https://github.com/os2display/display-admin-client/pull/279)